Define the interrupt vectors and reset handler.
The Cortex-M4 won't start without a defining the reset vector and a reset handler.

I copied these from the STMCube L4 templates and removed all the cruft in an attempt to demonstrate the minimum required.

The reset handler does some initialisation and calls main. At the moment there aren't are variables to initialise so this could well be wrong.