Turn on compiler optimisation.
Part of the reason for this investigation is to see how fast I can switch a GPIO, optimisation will help ensure we're going as fast as possible.

I think without optimisation the LL function calls are not being inlined despite having the 'inline' modifier.

With this change the GPIO is toggled at 6.6 MHz and the waveform is degrading because the GPIO can't switch fast enough.