Specify archecture when assembling and linking.
I now believe it is correct to tell the assembler and linker about the archecture.

The assembly files already included directives to do this so I guess this won't make much difference.

The linker was assuming the wrong archecture, "'-mcpu=arm7tdmi' '-mfloat-abi=soft' '-marm' '-march=armv4t'", which I could see by adding the "-v" option. Obviously it was working but it probably is better to have it correct!